collision logic added

+++ core/src/main/java/io/github/AngryBird/ChiefPig.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.physics.box2d.World;
 
 public class ChiefPig extends BasePig{
     Texture img=new Texture("chiefPig.png");
-    public ChiefPig(World world, int health, int x, int y) {
+    public ChiefPig(World world, int health, float x, float y) {
         super(world, health, x, y);
     }
     public void draw(SpriteBatch batch, float x, float y, float sizeX, float sizeY) {
