+++ core/src/main/java/io/github/AngryBird/Pig.java
@@ -8,7 +8,7 @@ import com.badlogic.gdx.physics.box2d.World;
 public class Pig extends BasePig{
     Texture img=new Texture("pig.png");
 
-    public Pig(World world, int health, int x, int y) {
+    public Pig(World world, int health, float x, float y) {
         super(world, health, x, y);
     }
     public void draw(SpriteBatch batch, float x, float y, float sizeX, float sizeY) {
